+++ Assets/Scripts/Game/NarrativeGenerator/Quests/ProfileQuests/Achievement/ExchangeQuestSo.cs
@@ -16,7 +16,7 @@ namespace Game.NarrativeGenerator.Quests.QuestGrammarTerminals
 
         public override string SymbolType => Constants.ExchangeQuest;
 
-        public override Dictionary<string, Func<int,int>> NextSymbolChances
+        public override Dictionary<string, Func<int,float>> NextSymbolChances
         {
             get => _nextSymbolChances;
             set => _nextSymbolChances = value;
