+++ Assets/Scripts/Game/NarrativeGenerator/Quests/ProfileQuests/Mastery/KillQuestSo.cs
@@ -16,7 +16,7 @@ namespace Game.NarrativeGenerator.Quests.QuestGrammarTerminals
         public Dictionary<float, int> EnemiesToKillByFitness { get; set; }
         public override string SymbolType => Constants.KillQuest;
 
-        public override Dictionary<string, Func<int,int>> NextSymbolChances
+        public override Dictionary<string, Func<int,float>> NextSymbolChances
         {
             get => _nextSymbolChances;
             set => _nextSymbolChances = value;
